+++ src/js/components/one-card-btn.js
@@ -1,17 +1,27 @@
 import LibraryBtn from './library-btn';
 import libraryType from './library-type';
 
-function oneCardOverlayBtn(movieObj) {
-  const watchBtn = new LibraryBtn({
-    element: document.querySelector('[data-action="add-to-watched_one-card"]'),
-    movieObj: movieObj,
-    type: libraryType.WATCHED,
-  });
+function createCardOverlay(movieObjs) {
+  if (pageState.isHome) {
+    const cards = document.querySelectorAll('.one-card_overlay');
 
-  const queueBtn = new LibraryBtn({
-    element: document.querySelector('[data-action="add-to-queue_one-card"]'),
-    movieObj: movieObj,
-    type: libraryType.QUEUE,
-  });
+    cards.forEach((card, index) => {
+      const watchBtnOne = new LibraryBtn({
+        element: card.querySelector('[data-action="add-to-watched_one-card"]'),
+        movieObj: movieObjs[index],
+        type: libraryType.WATCHED,
+      });
+
+      const queueBtnOne = new LibraryBtn({
+        element: card.querySelector('[data-action="add-to-queue_one-card"]'),
+        movieObj: movieObjs[index],
+        type: libraryType.QUEUE,
+      });
+    });
+  } else {
+    const oneCardBtn = document.querySelectorAll('.one-card_overlay');
+    [...oneCardBtn].map(element => element.classList.add('visually-hidden'));
+  }
 }
-export { oneCardOverlayBtn };
+
+export { createCardOverlay };
